+++ src/containers/Search/SearchList/SearchList.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styles from './SearchList.module.css';
 import { ReactComponent as SearchIcon } from '../../../assets/images/icons/Search/Dark.svg';
 import { List } from '../../List/List';
-import { COLLECTION_QUERY, COLLECTION_QUERY_COUNT } from '../../../graphql/queries/Collection';
+import { SEARCH_QUERY, SEARCH_QUERY_COUNT } from '../../../graphql/queries/Collection';
 import { DELETE_SEARCH } from '../../../graphql/mutations/Search';
 
 export interface SearchListProps {}
@@ -22,8 +22,8 @@ const columnStyles = [styles.Shortcode, styles.Name, styles.Actions];
 const searchIcon = <SearchIcon className={styles.Icon} />;
 
 const queries = {
-  countQuery: COLLECTION_QUERY_COUNT,
-  filterItemsQuery: COLLECTION_QUERY,
+  countQuery: SEARCH_QUERY_COUNT,
+  filterItemsQuery: SEARCH_QUERY,
   deleteItemQuery: DELETE_SEARCH,
 };
